Set Net::HTTP timeouts in Request's initializer
Timeout configuration is a client concern, but currently the
configuration is buried deep in internals of the library. So, as a first
step, we can move the configuration of the Request to the initializer.
We can leverage kwargs to provide default values for timeouts and make
them configurable from outside at the same time.